+++ tests/smoke-test.ts
@@ -1540,6 +1540,58 @@ const cases: {
       const billing = await client.billing.fetchUpcomingInvoice();
     },
   },
+
+  {
+    operation: 'list',
+    method: 'GET',
+    path: '/video/sources/live',
+    run: async () => {
+      const liveStreamWorkspace = await client.liveStreamWorkspaces.list();
+    },
+  },
+
+  {
+    operation: 'create',
+    method: 'POST',
+    path: '/video/sources/live',
+    run: async () => {
+      const liveStreamWorkspace = await client.liveStreamWorkspaces.create({
+        name: '',
+      });
+    },
+  },
+
+  {
+    operation: 'update',
+    method: 'POST',
+    path: '/video/sources/live/{live_workspace_id}',
+    label: 'required params',
+    run: async () => {
+      const liveStreamWorkspace = await client.liveStreamWorkspaces.update('liveWorkspaceId');
+    },
+  },
+
+  {
+    operation: 'update',
+    method: 'POST',
+    path: '/video/sources/live/{live_workspace_id}',
+    label: 'all params',
+    run: async () => {
+      const liveStreamWorkspace = await client.liveStreamWorkspaces.update('liveWorkspaceId', {
+        name: 'live-stream-collections',
+        video_source_id: '67bea1d66ca0059a95bf7de9',
+      });
+    },
+  },
+
+  {
+    operation: 'delete',
+    method: 'DELETE',
+    path: '/video/sources/live/{live_workspace_id}',
+    run: async () => {
+      const liveStreamWorkspace = await client.liveStreamWorkspaces.delete('liveWorkspaceId');
+    },
+  },
 ];
 
 /**